requirements.txt
Pillow==10.1.0 
pandas==2.1.3 